refactor(LanguageBar): port to Rust

+++ cpp/SampleIME/cbindgen/ime.h
@@ -7,6 +7,6 @@ extern "C" {
 
 extern HINSTANCE DLL_INSTANCE;
 
-HICON get_icon(int32_t desired_size, uint16_t index);
+HICON get_icon(int32_t desired_size, uint32_t index);
 
 } // extern "C"
